dt-bindings: fix up support for USB function
We can't have 2 nodes with the same exact name "usb@e8010000".

But, we need 2 nodes for the same hardware because we need one for host
and another for gadget so the user can choose. They have to be different
because the compatible name has to be different. Since usb@ is being used
upstream for gadget, we'll just rename the host nodes for our own BSP.

Also adding the nodes used for testing while we're at it.